+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/Input.scala
@@ -17,7 +17,6 @@ package com.intel.analytics.bigdl.nn
 
 import com.intel.analytics.bigdl.nn.Graph.ModuleNode
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, TensorModule}
-import com.intel.analytics.bigdl.nn.keras.Shape
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.Node
@@ -37,22 +36,8 @@ import scala.reflect.ClassTag
  * @tparam T The numeric type in the criterion, usually which are [[Float]] or [[Double]]
  */
 @SerialVersionUID(- 8525406230282608924L)
-class Input[T: ClassTag](val inputShape: Shape)(implicit ev: TensorNumeric[T])
+class Input[T: ClassTag]()(implicit ev: TensorNumeric[T])
   extends AbstractModule[Activity, Activity, T] {
-
-  /**
-   * There's no batch dim in the inputShape which just represent a sample record.
-   */
-  override def getInputShape(): Shape = {
-    inputShape
-  }
-
-  override def getOutputShape(): Shape = {
-    inputShape
-  }
-
-  override def computeOutputShape(inputShape: Shape): Shape = inputShape
-
   override def updateOutput(input: Activity): Activity = {
     output = input
     output
@@ -70,24 +55,11 @@ class Input[T: ClassTag](val inputShape: Shape)(implicit ev: TensorNumeric[T])
 }
 
 object Input {
-  def apply[T: ClassTag](name : String = null,
-    inputShape: Array[Int] = null)(implicit ev: TensorNumeric[T]): ModuleNode[T] = {
-    val module = new Input(Shape(inputShape))
+  def apply[T: ClassTag](name : String = null)(implicit ev: TensorNumeric[T]): ModuleNode[T] = {
+    val module = new Input()
     if (name != null) {
       module.setName(name)
     }
     new Node(module.asInstanceOf[AbstractModule[Activity, Activity, T]])
-  }
-}
-
-object InputLayer {
-  def apply[T: ClassTag](name : String = null,
-                         inputShape: Shape = null)(implicit ev: TensorNumeric[T])
-  : Input[T] = {
-    val module = new Input(inputShape)
-    if (name != null) {
-      module.setName(name)
-    }
-    module
   }
 }
