+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/keras/Topology.scala
@@ -21,7 +21,7 @@ import com.intel.analytics.bigdl.nn.Graph.ModuleNode
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
 import com.intel.analytics.bigdl.nn.{Graph, StaticGraph, Sequential => TSequential}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
-import com.intel.analytics.bigdl.utils.Node
+import com.intel.analytics.bigdl.utils.{Node, Shape}
 
 import scala.collection.mutable.ArrayBuffer
 import scala.reflect.ClassTag
