+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/keras/nn/KerasStyleSpec.scala
@@ -17,7 +17,7 @@
 package com.intel.analytics.bigdl.keras.nn
 
 import com.intel.analytics.bigdl.nn.keras.{Dense, Input, Model, Sequential => KSequential}
-import com.intel.analytics.bigdl.nn.{Input => TInput, Sequential => TSequential, _}
+import com.intel.analytics.bigdl.nn.{Sequential => TSequential, _}
 import com.intel.analytics.bigdl.numeric.NumericFloat
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.BigDLSpecHelper
