+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/utils/TestUtils.scala
@@ -18,9 +18,8 @@ package com.intel.analytics.bigdl.utils
 
 import java.util.concurrent.atomic.AtomicInteger
 
-import com.intel.analytics.bigdl.nn.{InputLayer, Sequential}
+import com.intel.analytics.bigdl.nn.keras.{InputLayer, Sequential => KSequential}
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity, TensorModule}
-import com.intel.analytics.bigdl.nn.keras.Shape
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import org.scalatest.exceptions.TestCanceledException
@@ -35,7 +34,7 @@ object TestUtils {
   def compareOutputShape(layer: AbstractModule[Activity, Activity, Float],
                             inputShape: Shape): Boolean = {
     val inputData = Tensor[Float](Array(2) ++ inputShape.toSingle()).randn()
-    val seq = Sequential[Float]()
+    val seq = KSequential[Float]()
     seq.add(InputLayer[Float](inputShape = inputShape))
     seq.add(layer)
     val calcOutputShape = seq.getOutputShape().toSingle()
